Improved error handling when download map offline

+++ src/app/pages/offline-map/offline-map.page.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { OfflineMapService } from '../../core/services/offline-map/offline-map.service';
 import { OfflineMapPackage } from '../../core/services/offline-map/offline-map.model';
 import { HelperService } from '../../core/services/helpers/helper.service';
-import { ActionSheetController, ModalController, Platform } from '@ionic/angular';
+import { ActionSheetController, AlertController, ModalController, Platform } from '@ionic/angular';
 import { ActionSheetButton } from '@ionic/core';
 import { BehaviorSubject, combineLatest, from, Observable, Subject } from 'rxjs';
 import { bufferTime, debounceTime, filter, map, shareReplay, switchMap, withLatestFrom } from 'rxjs/operators';
@@ -13,13 +13,16 @@ import { HttpClient } from '@angular/common/http';
 import { OfflinePackageModalComponent } from './offline-package-modal/offline-package-modal.component';
 import { CompoundPackageMetadata, FeatureProperties, PackageMetadata } from './metadata.model';
 import { Polygon, Feature } from 'geojson';
+import { TranslateService } from '@ngx-translate/core';
 
 const PACKAGE_INDEX_URL = "https://offlinemap.blob.core.windows.net/packages/packageIndex.json";
 const documentStyle = getComputedStyle(document.body);
 const defaultTileStyle = {
   color: documentStyle.getPropertyValue("--ion-color-primary")
 };
-
+const errorTileStyle = {
+  color: documentStyle.getPropertyValue("--ion-color-danger")
+};
 
 
 interface PackageIndex {
@@ -37,6 +40,7 @@ interface PackageIndex {
 })
 export class OfflineMapPage {
   packages$: Observable<OfflineMapPackage[]>;
+  downloadAndUnzipProgress$: Observable<OfflineMapPackage[]>;
   private packageIndex$: Observable<PackageIndex>;
   selectedPackage: CompoundPackageMetadata = null;
   packageMap = new Map<string, CompoundPackageMetadata>();
@@ -46,6 +50,7 @@ export class OfflineMapPage {
   // Could not get the click handler to only emit once per click, so wrapped this in a subject
   showModal = new Subject<Feature<Polygon, FeatureProperties>>();
   isZooming = new BehaviorSubject<boolean>(false);
+  featureMap = new Map<string, Feature<Polygon, FeatureProperties>>();
 
   constructor(
     private helperService: HelperService,
@@ -53,10 +58,12 @@ export class OfflineMapPage {
     private platform: Platform,
     private modalController: ModalController,
     private offlineMapService: OfflineMapService,
+    private alertController: AlertController,
+    private translateService: TranslateService,
     http: HttpClient,
   ) {
     this.packageIndex$ = http.get<PackageIndex>(PACKAGE_INDEX_URL).pipe(shareReplay());
-
+    this.downloadAndUnzipProgress$ = this.offlineMapService.downloadAndUnzipProgress$;
     this.packages$ = combineLatest([
       this.offlineMapService.packages$,
       this.offlineMapService.downloadAndUnzipProgress$
@@ -77,6 +84,9 @@ export class OfflineMapPage {
       this.packageIndex$,
       this.packages$
     ]).subscribe(([packageIndex, packages]) => {
+
+      // TODO: Vil ikke dette skje litt vel ofte?? Hver gang en pakke som nedlastes får en endring på progress, så trigges dette...
+
       try {
         this.packageMap = new Map<string, CompoundPackageMetadata>();
         packageIndex.statensKartverk.forEach(p => {
@@ -96,19 +106,22 @@ export class OfflineMapPage {
           }
         }
   
-        const installedPackages = new Map<string, OfflineMapPackage>();
-        packages.forEach(installedPackage => {
-          const maps = Object.values(installedPackage.maps);
-          if (maps.length > 0) {
-            const { x, y, z } = Object.values(installedPackage.maps)[0].rootTile;
-            const id = this.getFeatureId(x, y, z);
-            installedPackages.set(id, installedPackage);
-          }
-        });
+        // const installedPackages = new Map<string, OfflineMapPackage>();
+        // packages.forEach(installedPackage => {
+        //   const maps = Object.values(installedPackage.maps);
+        //   if (maps.length > 0) {
+        //     const { x, y, z } = Object.values(installedPackage.maps)[0].rootTile;
+        //     const id = this.getFeatureId(x, y, z);
+        //     installedPackages.set(id, installedPackage);
+        //   }
+        // });
+        const installedPackages = new Map(packages.map((p) => [this.getFeatureIdForPackage(p), p]));
   
         if (this.tilesLayer) {
           map.removeLayer(this.tilesLayer);
         }
+
+        this.featureMap = new Map<string, Feature<Polygon, FeatureProperties>>();
   
         this.tilesLayer = new L.GeoJSON(tiles as GeoJsonObject, {
           filter: (feature) => {
@@ -121,6 +134,7 @@ export class OfflineMapPage {
             // }
 
             if (this.packageMap.has(feature.id as string)) {
+              this.featureMap.set(feature.id as string, feature);
               layer.on("click", () => {
                 this.showModal.next(feature);
                 // this.showPackageModal(feature);
@@ -131,10 +145,17 @@ export class OfflineMapPage {
         });
   
         this.tilesLayer.setStyle(feature => {
-          if (installedPackages.has(feature.id as string)) {
+          const installedPackage = installedPackages.get(feature.id as string); 
+          if (installedPackage != null) {
+            if(installedPackage.error) {
+              return {
+                ...errorTileStyle,
+                fillOpacity:  0.8,
+              }
+            }
             return {
               ...defaultTileStyle,
-              fillOpacity: 0.8,
+              fillOpacity: installedPackage.downloadComplete ?  0.8 : 0.5,
             }
           }
   
@@ -175,6 +196,22 @@ export class OfflineMapPage {
     return `(${x}, ${y}, ${z})`;
   }
 
+  private getFeatureIdFromXYZ(xyz: number[]) {
+    const [x, y, z] = xyz;
+    return this.getFeatureId(x, y, z);
+  }
+
+  private getFeatureIdForPackage(map: OfflineMapPackage): string {
+    if(map.compoundPackageMetadata) {
+      return this.getFeatureIdFromXYZ(map.compoundPackageMetadata.getXYZ());
+    }
+    if(map.maps['statensKartverk']) {
+      const rootTile = map.maps['statensKartverk'].rootTile;
+      return this.getFeatureId(rootTile.x, rootTile.y, rootTile.z);
+    }
+    return '';
+  }
+
   async showPackageModal(feature: Feature<Polygon, FeatureProperties>) {
     const [x, y, z] = feature.properties.xyz;
     const modal = await this.modalController.create({
@@ -191,6 +228,13 @@ export class OfflineMapPage {
     await modal.present();
   }
 
+  showPackageModalForPackage(map: OfflineMapPackage) {
+    const feature = this.featureMap.get(this.getFeatureIdFromXYZ(map.compoundPackageMetadata.getXYZ()));
+    if(feature) {
+      this.showPackageModal(feature);
+    }
+  }
+
   humanReadableByteSize(bytes: number): string {
     if (isNaN(bytes)) {
       return '';
@@ -202,13 +246,38 @@ export class OfflineMapPage {
     return Math.round((map.progress ? map.progress.percentage : 0) * 100);
   }
 
-  deleteMap(map: OfflineMapPackage): Promise<void> {
-    return this.offlineMapService.removeMapPackageByName(map.name);
+  async cancelOrDelete(map: OfflineMapPackage, event: Event) {
+    event.stopPropagation();
+    if (this.isDownloaded(map)) {
+      // TODO: Are you su
+      const toTranslate = ['DIALOGS.ARE_YOU_SURE', 'DIALOGS.CANCEL', 'DIALOGS.OK'];
+      const translations = await this.translateService
+        .get(toTranslate)
+        .toPromise();
+      const alert = await this.alertController.create({
+        header: translations['DIALOGS.ARE_YOU_SURE'],
+        message: translations['DIALOGS.ARE_YOU_SURE'],
+        buttons: [
+          {
+            text: translations['DIALOGS.CANCEL'],
+            role: 'cancel'
+          },
+          {
+            text: translations['DIALOGS.OK'],
+            handler: () => {
+              this.offlineMapService.removeMapPackageByName(map.name);
+            }
+          }
+        ]
+      });
+      alert.present();
+    }else{
+      this.offlineMapService.cancelDownloadPackage(map);
+    }
   }
 
-  async cancelUnzip(map: OfflineMapPackage): Promise<void> {
-    await this.offlineMapService.removeMapPackageByName(map.name);
-  }
+
+
 
   isDownloading(map: OfflineMapPackage): boolean {
     return map.downloadStart && !map.downloadComplete;
@@ -222,39 +291,39 @@ export class OfflineMapPage {
     this.selectedPackage = null;
   }
 
-  async presentActionSheet(map: OfflineMapPackage): Promise<void> {
-    const header = map.name;
-    const subHeader = this.humanReadableByteSize(map.size);
-    const buttons: ActionSheetButton[] = [];
-    if (this.isDownloaded(map)) {
-      buttons.push({
-        text: 'Delete',
-        role: 'destructive',
-        icon: 'trash',
-        handler: () => {
-          this.offlineMapService.removeMapPackageByName(map.name);
-        }
-      });
-    } else if (this.isDownloading(map)) {
-      buttons.push({
-        text: 'Avbryt',
-        icon: 'close',
-        handler: () => {
-          this.cancelUnzip(map);
-        }
-      });
-    }
+  // async presentActionSheet(map: OfflineMapPackage): Promise<void> {
+  //   const header = map.name;
+  //   const subHeader = this.humanReadableByteSize(map.size);
+  //   const buttons: ActionSheetButton[] = [];
+  //   if (this.isDownloaded(map)) {
+  //     buttons.push({
+  //       text: 'Delete',
+  //       role: 'destructive',
+  //       icon: 'trash',
+  //       handler: () => {
+  //         this.offlineMapService.removeMapPackageByName(map.name);
+  //       }
+  //     });
+  //   } else if (this.isDownloading(map)) {
+  //     buttons.push({
+  //       text: 'Avbryt',
+  //       icon: 'close',
+  //       handler: () => {
+  //         this.cancelUnzip(map);
+  //       }
+  //     });
+  //   }
 
-    buttons.push({
-      text: 'Lukk',
-      role: 'cancel'
-    });
+  //   buttons.push({
+  //     text: 'Lukk',
+  //     role: 'cancel'
+  //   });
 
-    const actionSheet = await this.actionSheetController.create({
-      header,
-      subHeader,
-      buttons
-    });
-    await actionSheet.present();
-  }
+  //   const actionSheet = await this.actionSheetController.create({
+  //     header,
+  //     subHeader,
+  //     buttons
+  //   });
+  //   await actionSheet.present();
+  // }
 }
